requirements.txt
PySimpleGUI
numpy
Pillow
matplotlib
scipy
pandas
